+++ data/src/main/java/com/example/habits/data/database/HabitDatabase.kt
@@ -3,7 +3,7 @@ package com.example.habits.data.database
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.example.habits.data.model_vo.HabitItem
+import com.example.habits.domain.models.model_vo.HabitItem
 
 @Database(entities = [HabitItem::class], version = 1)
 @TypeConverters(HabitTypeConverter::class, HabitTypeConverter::class, ListTypeConverter::class)
