+++ data/src/main/java/com/example/habits/data/database/HabitTypeConverter.kt
@@ -1,7 +1,7 @@
 package com.example.habits.data.database
 
 import androidx.room.TypeConverter
-import com.example.habits.data.model_vo.HabitType
+import com.example.habits.domain.models.model_vo.HabitType
 
 class HabitTypeConverter {
