+++ data/src/main/java/com/example/habits/data/network/HabitApiInterface.kt
@@ -1,8 +1,8 @@
 package com.example.habits.data.network
 
-import com.example.habits.data.model_dto.HabitDoneDto
-import com.example.habits.data.model_dto.HabitDto
-import com.example.habits.data.model_dto.HabitUidDto
+import com.example.habits.domain.models.model_dto.HabitDoneDto
+import com.example.habits.domain.models.model_dto.HabitDto
+import com.example.habits.domain.models.model_dto.HabitUidDto
 import retrofit2.http.*
 
 interface HabitApiInterface {
