+++ domain/src/main/java/com/example/habits/domain/models/model_dto/HabitDoneDto.kt
@@ -1,4 +1,4 @@
-package com.example.habits.data.model_dto
+package com.example.habits.domain.models.model_dto
 
 import kotlinx.serialization.SerialName
