+++ domain/src/main/java/com/example/habits/domain/models/model_vo/HabitType.kt
@@ -1,4 +1,4 @@
-package com.example.habits.data.model_vo
+package com.example.habits.domain.models.model_vo
 
 import android.os.Parcelable
 import androidx.annotation.StringRes
