+++ presentation/src/main/java/com/example/habits/presentation/adapter/HabitViewPagerAdapter.kt
@@ -2,7 +2,7 @@ package com.example.habits.presentation.adapter
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.example.habits.data.model_vo.HabitType
+import com.example.habits.domain.models.model_vo.HabitType
 import com.example.habits.presentation.list.HabitsListFragment
 
 class HabitViewPagerAdapter(fragment: Fragment): FragmentStateAdapter(fragment) {
