+++ presentation/src/main/java/com/example/habits/presentation/creator/HabitCreatorFragment.kt
@@ -16,12 +16,12 @@ import by.kirich1409.viewbindingdelegate.viewBinding
 import com.example.habits.App
 import com.example.habits.R
 import com.example.habits.data.colorpicker.ColorPicker
-import com.example.habits.data.extension.factory
-import com.example.habits.data.extension.getBackgroundColor
-import com.example.habits.data.model_vo.HabitType
-import com.example.habits.data.extension.hideKeyboard
-import com.example.habits.data.model_vo.HabitItem
-import com.example.habits.data.model_vo.PriorityType
+import com.example.habits.presentation.extension.factory
+import com.example.habits.presentation.extension.getBackgroundColor
+import com.example.habits.domain.models.model_vo.HabitType
+import com.example.habits.presentation.extension.hideKeyboard
+import com.example.habits.domain.models.model_vo.HabitItem
+import com.example.habits.domain.models.model_vo.PriorityType
 import com.example.habits.data.network.HabitApiClient
 import com.example.habits.databinding.FragmentHabitCreatorBinding
 import com.google.android.material.snackbar.Snackbar
