+++ presentation/src/main/java/com/example/habits/presentation/creator/HabitCreatorViewModel.kt
@@ -2,8 +2,8 @@ package com.example.habits.presentation.creator
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.habits.data.model_dto.HabitUidDto
-import com.example.habits.data.model_vo.HabitItem
+import com.example.habits.domain.models.model_dto.HabitUidDto
+import com.example.habits.domain.models.model_vo.HabitItem
 import com.example.habits.domain.usecase.HabitCreatorUseCase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
