+++ presentation/src/main/java/com/example/habits/presentation/extension/View.kt
@@ -1,4 +1,4 @@
-package com.example.habits.data.extension
+package com.example.habits.presentation.extension
 
 import android.graphics.Color
 import android.view.View
