+++ presentation/src/main/java/com/example/habits/presentation/list/HabitsListFragment.kt
@@ -19,12 +19,12 @@ import com.example.habits.App
 import com.example.habits.R
 import com.example.habits.presentation.adapter.HabitAdapter
 import com.example.habits.databinding.FragmentHabitsListBinding
-import com.example.habits.data.model_vo.HabitType
-import com.example.habits.data.extension.addToggleToNavigationDrawer
-import com.example.habits.data.extension.afterTextChanged
-import com.example.habits.data.extension.factory
-import com.example.habits.data.model_dto.HabitDoneDto
-import com.example.habits.data.model_vo.HabitItem
+import com.example.habits.domain.models.model_vo.HabitType
+import com.example.habits.presentation.extension.addToggleToNavigationDrawer
+import com.example.habits.presentation.extension.afterTextChanged
+import com.example.habits.presentation.extension.factory
+import com.example.habits.domain.models.model_dto.HabitDoneDto
+import com.example.habits.domain.models.model_vo.HabitItem
 import com.example.habits.data.network.HabitApiClient
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import kotlinx.coroutines.*
